mvc project again

+++ _0MVC_BOARD/src/main/webapp/deleteForm.jsp
@@ -1,30 +1,33 @@
+<%@page import="dto.Board"%>
 <%@page import="dao.BoardDao"%>
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
-
-<%@ page import="java.sql.*"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 
 <%
-// 지정된 글 번호 얻기
+// 세션에서 로그인된 사용자 정보 가져오기
+int loggedInMemberNo = (int) session.getAttribute("MEMBERNO");
+
+// 게시글 번호 받아오기
 int num = Integer.parseInt(request.getParameter("num"));
 
-String memberId = (String) session.getAttribute("MEMBERID");
-if (memberId == null) {
-	response.sendRedirect("list.do");
-}
+// 게시글 작성자의 memberNo 가져오기
 BoardDao dao = BoardDao.getInstance();
-dao.delete(num);
+Board board = dao.selectOne(num, false); // false를 전달하여 작성자 정보를 함께 조회하지 않음
+
+// 작성자 정보와 로그인한 사용자 정보가 일치하는지 확인
+if (board != null && loggedInMemberNo == board.getMemberno()) {
+    dao.delete(num); // 삭제 수행
 %>
-
-        <script>
-            alert('삭제가 되었습니다.');
-            location.href="list.do";
-        </script>
-
+    <script>
+        alert('삭제가 되었습니다.');
+        location.href="list.do";
+    </script>
 <%
-try {
-    response.sendRedirect("list.do");
-} catch (Exception e) {
-    e.printStackTrace();
+} else {
+%>
+    <script>
+        alert('본인의 글만 삭제할 수 있습니다.');
+        location.href="list.do";
+    </script>
+<%
 }
 %>
